+++ src/actions/fetchAction.js
@@ -1,5 +1,3 @@
-import store from "../store";
-
 export const fetch_post = () => {
   return {
     type: "FETCH_USER"
@@ -16,20 +14,5 @@ export const receive_post = post => {
 export const receive_error = () => {
   return {
     type: "RECEIVE_ERROR"
-  };
-};
-
-export const thunk_action_creator = username => {
-  const user = username.replace(/\s/g, "");
-  store.dispatch(fetch_post());
-  return function (dispatch, getState) {
-    return fetch(`https://api.github.com/users/${user}`)
-      .then(data => data.json())
-      .then(data => {
-        if (data.message === "Not Found") {
-          throw new Error("No such user found!!");
-        } else dispatch(receive_post(data));
-      })
-      .catch(err => dispatch(receive_error()));
   };
 };
